Add a note on ConfigMap data size
The [Kubernetes documentation](https://kubernetes.io/docs/concepts/configuration/configmap/#motivation) has this to say:
A ConfigMap is not designed to hold large chunks of data. The data stored in a ConfigMap cannot exceed 1 MiB.
If you need to store settings that are larger than this limit, you may want to consider mounting a volume or use
a separate database or file service.