Update manifest post release v0.28.0
1. Add `sveltosctl generate kubeconfig` command. Pointing to the cluster
to be managed, this command generates a kubeconfig which can be later used
to register the cluster with Sveltos
2. The `sveltosctl register cluster` command has been extended and can now
be passed a set of labels. Those labels will be added to the SveltosCluster
instance generated by the register command.
3. The `sveltosctl register cluster` has been extended to generate a kubeconfig
and register a cluster. If kubeconfig has multiple contexts, and the default
context points to the management cluster, use the __--fleet-cluster-context__ option.
Set it to the name of the context that points to the cluster you want to register.
Run the command like this:

```
sveltosctl register cluster --namespace=<namespace> --cluster=<cluster name> --fleet-cluster-context=<context name> --labels=key1=value1,key2=value2
```

To be merged once release v0.28.0 is ready