Update documentation post release v0.51.1
**Deploying Helm Charts with Flux and Sveltos: Added Example**

We have expanded the documentation to include a comprehensive
example demonstrating how to deploy Helm charts using a combination
of Flux and Sveltos.

**Implicit Dependency Resolution with `dependsOn`**

Sveltos offers a powerful feature for managing dependencies
between profiles through the `dependsOn` field. This allows
you to define the order in which resources are deployed,
ensuring that prerequisites are installed before dependents.